add documentation, fix compiler dependencies, refactoring
