+++ analyzer/src/steps/typing/exploration.rs
@@ -28,10 +28,9 @@ impl Exploration {
 
 /// Generates a diagnostic for an unknown type annotation.
 pub(super) fn diagnose_unknown_type(source: SourceId, segment: SourceSegment) -> Diagnostic {
-    Diagnostic::new(DiagnosticID::UnknownType, source, "Unknown type annotation")
-        .with_observation(
-            Observation::new(segment)
+    Diagnostic::new(DiagnosticID::UnknownType, source, "Unknown type annotation").with_observation(
+        Observation::new(segment)
             .with_help("Not found in scope")
-                .with_tag(ObservationTag::InFault)
-        )
+            .with_tag(ObservationTag::InFault),
+    )
 }
