+++ cli/src/repl.rs
@@ -1,4 +1,5 @@
-use crate::cli::{display_bytecode, display_exprs, display_tokens, execute, Configuration};
+use crate::cli::{display_exprs, display_tokens, execute, Configuration};
+use crate::display_bytecode::display_bytecode;
 use crate::formatted_diagnostic::render_diagnostic;
 use crate::formatted_parse_error::render_parse_error;
 use analyzer::diagnostic::Diagnostic;
