+++ compiler/Cargo.toml
@@ -9,5 +9,4 @@ edition = "2021"
 analyzer = { path = "../analyzer" }
 ast = { path = "../ast" }
 indexmap = "1.9.3"
-enum-assoc = "1.0.0"
-
+num_enum = "0.6.1"
