+++ compiler/src/bytecode.rs
@@ -1,5 +1,4 @@
-use crate::bytecode::Opcode::*;
-use enum_assoc::Assoc;
+use num_enum::TryFromPrimitive;
 use std::mem::size_of;
 
 /// Holds the currently generated bytecode.
@@ -26,12 +25,12 @@ impl Bytecode {
     }
 
     pub fn emit_get_local(&mut self, identifier: u8) {
-        self.emit_code(GetLocal);
+        self.emit_code(Opcode::GetLocal);
         self.bytes.push(identifier);
     }
 
     pub fn emit_int(&mut self, constant: i64) {
-        self.emit_code(PushInt);
+        self.emit_code(Opcode::PushInt);
         self.bytes.extend(constant.to_be_bytes());
     }
 
@@ -41,12 +40,12 @@ impl Bytecode {
     }
 
     pub fn emit_float(&mut self, constant: f64) {
-        self.emit_code(PushFloat);
+        self.emit_code(Opcode::PushFloat);
         self.bytes.extend(constant.to_be_bytes());
     }
 
     pub fn emit_string_constant_ref(&mut self, constant_ref: usize) {
-        self.emit_code(PushString);
+        self.emit_code(Opcode::PushString);
         self.bytes.extend(constant_ref.to_be_bytes());
     }
 
@@ -100,83 +99,33 @@ impl Bytecode {
 }
 
 /// see vm's `Opcode` enum for more details
+#[derive(TryFromPrimitive)]
 #[repr(u8)]
-#[derive(Assoc)]
-#[func(pub fn mnemonic(&self) -> &'static str)]
 pub enum Opcode {
-    #[assoc(mnemonic = "pshi")]
     PushInt,
-    #[assoc(mnemonic = "pshb")]
     PushByte,
-    #[assoc(mnemonic = "pshf")]
     PushFloat,
-    #[assoc(mnemonic = "pshs")]
     PushString,
-    #[assoc(mnemonic = "get")]
     GetLocal,
-    #[assoc(mnemonic = "set")]
     SetLocal,
-    #[assoc(mnemonic = "spawn")]
     Spawn,
 
-    #[assoc(mnemonic = "pop")]
     PopByte,
-    #[assoc(mnemonic = "pop8")]
     PopQWord,
 
-    #[assoc(mnemonic = "ifjmp")]
     IfJump,
-    #[assoc(mnemonic = "ifnjmp")]
     IfNotJump,
-    #[assoc(mnemonic = "jmp")]
     Jump,
 
-    #[assoc(mnemonic = "i2s")]
     ConvertIntToStr,
-    #[assoc(mnemonic = "f2s")]
     ConvertFloatToStr,
-    #[assoc(mnemonic = "i2b")]
     ConvertIntToByte,
-    #[assoc(mnemonic = "cct")]
     Concat,
 
-    #[assoc(mnemonic = "bxor")]
     BXor,
-    #[assoc(mnemonic = "iadd")]
     IntAdd,
-    #[assoc(mnemonic = "isub")]
     IntSub,
-    #[assoc(mnemonic = "imul")]
     IntMul,
-    #[assoc(mnemonic = "idiv")]
     IntDiv,
-    #[assoc(mnemonic = "imod")]
     IntMod,
-}
-
-impl TryFrom<u8> for Opcode {
-    type Error = ();
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        let opcode = match value {
-            0 => PushInt,
-            1 => PushFloat,
-            2 => PushString,
-            3 => GetLocal,
-            4 => SetLocal,
-            5 => Spawn,
-
-            6 => PopByte,
-            7 => PopQWord,
-
-            8 => IfJump,
-            9 => IfNotJump,
-            10 => Jump,
-
-            11 => ConvertIntToStr,
-            12 => ConvertFloatToStr,
-            _ => return Err(()),
-        };
-        Ok(opcode)
-    }
 }
